backend/requirements.txt
# Core Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# AI & ML
google-generativeai>=0.4.0

# Vector Database
qdrant-client>=1.7.1

# Job Queue
redis>=5.0.1
rq>=1.16.0

# Video Processing
ffmpeg-python>=0.2.0

# Image Processing
Pillow>=10.2.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.3
pydantic-settings>=2.1.0
loguru>=0.7.2
httpx>=0.26.0
tenacity>=8.2.3

# Testing
pytest>=7.4.4
pytest-asyncio>=0.23.3
